eom-hub/backend/requirements.txt
eel
pyyaml
